Add support for deep link facet hack
This includes an Ebsco provided JavaScript to enable sending in URL
parameters that EDS UI doesn't natively support, such as being able to
apply facets to a search.

Initially, we are using this to to pass in a comma separated list of
facets to apply to a search to match the facets we apply to create our
bento results.

We hope this is only needed temporarily as it is very hacky and a true
native solution should be provided by Ebsco.